Move logic to EF relational project
Resolve: #29
