+++ easydata.net/src/EasyData.Core/MetaDescriptors/EntityMetadataDescriptor.cs
@@ -10,12 +10,12 @@ namespace EasyData.MetaDescriptors
     /// <summary>
     /// Contains metadata information about entity.
     /// </summary>
-    public class EntityMetadataDescriptor<TEntity>: IEntityMetadataDescriptor
+    public class EntityMetadataDescriptor
     {
         /// <summary>
-        /// Set of entity properties metadata.
+        /// Set of entity attributes metadata.
         /// </summary>
-        public HashSet<EntityPropertyMetadataDescriptor> MetadataProperties { get; } = new HashSet<EntityPropertyMetadataDescriptor>(new PropertyComparer());
+        public HashSet<EntityAttributeMetadataDescriptor> MetadataAttributes { get; } = new HashSet<EntityAttributeMetadataDescriptor>(new AttributeComparer());
 
         /// <summary>
         /// Name to use for this entity in the UI.
@@ -23,7 +23,7 @@ namespace EasyData.MetaDescriptors
         public string DisplayName { get; set; }
 
         /// <summary>
-        /// Plural name to use for this property in the UI.
+        /// Plural name to use in the UI.
         /// </summary>
         public string DisplayNamePlural { get; set; }
 
@@ -40,82 +40,14 @@ namespace EasyData.MetaDescriptors
         /// <summary>
         /// Entity type.
         /// </summary>
-        public Type ClrType => typeof(TEntity);
+        public Type ClrType { get; set; }
 
         /// <summary>
-        /// Set entity display name.
+        /// Compare entity attributes metadata.
         /// </summary>
-        /// <param name="displayName">Name to set.</param>
-        /// <returns>Current instance of the class.</returns>
-        public EntityMetadataDescriptor<TEntity> SetDisplayName(string displayName)
+        private class AttributeComparer : IEqualityComparer<EntityAttributeMetadataDescriptor>
         {
-            DisplayName = displayName;
-            return this;
-        }
-
-        /// <summary>
-        /// Set availability for the entity.
-        /// </summary>
-        /// <param name="enabled">Enable or not.</param>
-        /// <returns>Current instance of the class.</returns>
-        public EntityMetadataDescriptor<TEntity> SetEnabled(bool enabled)
-        {
-            IsEnabled = enabled;
-            return this;
-        }
-
-        /// <summary>
-        /// Set entity plural display name.
-        /// </summary>
-        /// <param name="displayNamePlural">Name to set.</param>
-        /// <returns>Current instance of the class.</returns>
-        public EntityMetadataDescriptor<TEntity> SetDisplayNamePlural(string displayNamePlural)
-        {
-            DisplayNamePlural = displayNamePlural;
-            return this;
-        }
-
-        /// <summary>
-        /// Set entity description.
-        /// </summary>
-        /// <param name="description">Description to set.</param>
-        /// <returns>Current instance of the class.</returns>
-        public EntityMetadataDescriptor<TEntity> SetDescription(string description)
-        {
-            Description = description;
-            return this;
-        }
-
-        /// <summary>
-        /// Get entity property metadata descriptor.
-        /// </summary>
-        /// <param name="propertySelector">Property expression.</param>
-        /// <returns>Property metadata descriptor instance.</returns>
-        public EntityPropertyMetadataDescriptor Property(Expression<Func<TEntity, object>> propertySelector)
-        {
-            PropertyInfo property;
-            if (propertySelector.Body is MemberExpression expression) {
-                property = (PropertyInfo)expression.Member;
-            }
-            else {
-                var memberExpression = ((UnaryExpression)propertySelector.Body).Operand as MemberExpression;
-                property = (PropertyInfo)memberExpression.Member;
-            }
-
-            var propertyMetadataDescriptor = new EntityPropertyMetadataDescriptor
-            {
-                PropertyInfo = property
-            };
-            MetadataProperties.Add(propertyMetadataDescriptor);
-            return propertyMetadataDescriptor;
-        }
-
-        /// <summary>
-        /// Compare entity properties metadata.
-        /// </summary>
-        private class PropertyComparer : IEqualityComparer<EntityPropertyMetadataDescriptor>
-        {
-            public bool Equals(EntityPropertyMetadataDescriptor x, EntityPropertyMetadataDescriptor y)
+            public bool Equals(EntityAttributeMetadataDescriptor x, EntityAttributeMetadataDescriptor y)
             {
                 if (x == null || y == null) {
                     return false;
@@ -124,7 +56,7 @@ namespace EasyData.MetaDescriptors
                 return x.PropertyInfo.Equals(y.PropertyInfo);
             }
 
-            public int GetHashCode(EntityPropertyMetadataDescriptor obj)
+            public int GetHashCode(EntityAttributeMetadataDescriptor obj)
             {
                 return obj.PropertyInfo.GetHashCode();
             }
